add Traits/Resource/Pagination.php
