+++ src/Zendesk/API/Traits/Resource/FindAll.php
@@ -4,10 +4,11 @@ namespace Zendesk\API\Traits\Resource;
 
 use Zendesk\API\Exceptions\RouteException;
 use Zendesk\API\Traits\Utility\Pagination\CbpStrategy;
-use Zendesk\API\Traits\Utility\Pagination\PaginationIterator;
 
 trait FindAll
 {
+    use Pagination;
+
     /**
      * List all of this resource
      *
@@ -36,31 +37,5 @@ trait FindAll
             $route,
             $params
         );
-    }
-
-    // TODO: own trait
-    // TODO: page size 100
-    /**
-     * Usage:
-     * foreach ($ticketsIterator as $ticket) {
-     *     process($ticket)
-     * }
-     *
-     * @return PaginationIterator to fetch all pages.
-     */
-    public function iterator()
-    {
-        $strategyClass = $this->paginationStrategyClass();
-        $strategy = new $strategyClass($this, $this->resourcesRoot(), 2);
-        return new PaginationIterator($strategy);
-    }
-
-
-    private function paginationStrategyClass() {
-        return CbpStrategy::class;
-    }
-
-    protected function resourcesRoot() {
-        return $this->objectNamePlural;
     }
 }
